+++ scripts/verify-updater.sh
@@ -18,6 +18,8 @@ case "$bundle_identifier" in
         expected_feed='https://github.com/pdparchitect/noodle/releases/download/applet-latest/appcast.xml' ;;
     com.pdparchitect.noodle.browser|com.pdparchitect.noodle.browser.local)
         expected_feed='https://github.com/pdparchitect/noodle/releases/download/browser-latest/appcast.xml' ;;
+    com.pdparchitect.noodle.hub|com.pdparchitect.noodle.hub.local)
+        expected_feed='https://github.com/pdparchitect/noodle/releases/download/hub-latest/appcast.xml' ;;
 esac
 [[ "$(/usr/libexec/PlistBuddy -c 'Print :SUFeedURL' "$info")" == "$expected_feed" ]]
 [[ "$(/usr/libexec/PlistBuddy -c 'Print :SUPublicEDKey' "$info")" == '1ZT5NrPiDPaQ54iHGSI1a9JIn6kTrmjQvzZRBA9f/sk=' ]]